requirements.txt
pandas==2.1.4
numpy==1.26.3
scikit-learn==1.3.2
matplotlib==3.8.2
pyyaml==6.0.1
python-dateutil==2.8.2
